pyproject.toml
[tool.poetry]
name = "staticdistributor"
version = "0.1.0"
description = ""
authors = ["Анастасия <Any4rd@gmail.com>"]
readme = "README.md"
packages = [{include = "staticdistributor"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.1"
pydantic = "^1.10.2"
requests = "^2.28.1"
uvicorn = "^0.19.0"
python-dotenv = "^0.21.0"
asyncpg = "^0.27.0"
dependency-injector = "^4.40.0"
loguru = "^0.6.0"
aiopg = "^1.4.0"
python-multipart = "^0.0.5"
psycopg2 = "^2.9.5"
psycopg2-binary = "^2.9.5"
pyjwt = "^2.7.0"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
